+++ src/main/java/com/buttons/smarthome/repo/LandLordRepo.java
@@ -3,5 +3,5 @@ package com.buttons.smarthome.repo;
 import com.buttons.smarthome.models.LandLord;
 import org.springframework.data.repository.CrudRepository;
 
-public interface LandLordRepo extends CrudRepository<LandLord, Integer> {
+public interface LandLordRepo extends CrudRepository<LandLord, Long> {
 }
